update script (to be run after cleaning up emoji)
Makes sure the "build" directory catches all my changes...

If the build directory exists:

Deletes everything that appears in "cleaned" from the build directory.

Regardless of if the build directory exists:

Overwrites the "assets/noto-color-emoji-svg" folder with
my changes from sub-folder "cleaned"; So the changed SVGs
will all be copied freshly to the build directory before rebuilding.

Makes rebuilding the fonts faster, because I don't have to run
emoji I haven't touched through SVGO and Potrace multiple times.

(This commit adds the copy operation, so all changes in cleaned
overwrite the SVGs from noto-color-emoji-svg. Also adds
a conditional check to see if the build dir exists. Lastly,
adds comments to the script to explain what it does.)

+++ assets/noto-color-emoji-svg/cleaned/delete build.sh
@@ -1,6 +1,15 @@
 for i in `ls -d *.svg`
-do
-echo "removing $i from svg-bw and svg-color"
-rm ../../../build/svg-bw/$i
-rm ../../../build/svg-color/$i
-done
+
+# if the build folder exists
+if [ -d "../../../build" ]; then
+
+# delete all files with an SVG in cleaned from the build folder
+  do
+  echo "removing $i from svg-bw and svg-color"
+  rm ../../../build/svg-bw/$i
+  rm ../../../build/svg-color/$i
+  done
+fi
+
+# regardless, copy from cleaned to the main `assets/noto-nolor-emoji-svg` folder
+cp ./*.svg ../
